+++ backend/packages/Core/src/domain/GetDocumentGroupRequest.ts
@@ -1,4 +1,12 @@
 import { Credentials } from "./Credentials";
 
-
-export type GetDocumentGroupRequest = { credentials: Credentials; groupId: string; };
+/**
+ * Request object for retrieving a specific document group by its ID.
+ */
+export type GetDocumentGroupRequest = { 
+    /** Authentication credentials for the API request */
+    credentials: Credentials; 
+    
+    /** Unique identifier of the document group to retrieve */
+    groupId: string; 
+};
