+++ backend/packages/Core/src/domain/GetDocumentTypeSchemaRequest.ts
@@ -1,4 +1,12 @@
 import { Credentials } from "./Credentials";
 
-
-export type GetDocumentTypeSchemaRequest = { credentials: Credentials; documentTypeId: string; };
+/**
+ * Request object for retrieving the schema of a specific document type.
+ */
+export type GetDocumentTypeSchemaRequest = { 
+    /** Authentication credentials for the API request */
+    credentials: Credentials;
+    
+    /** Unique identifier of the document type whose schema is being requested */
+    documentTypeId: string; 
+};
