+++ backend/packages/server/routes/handlers/handleGetDocumentTypeSchema.ts
@@ -5,7 +5,7 @@ import { resolveSchemaService } from "../utils/resolveSchemaService";
 
 export const handleGetDocumentTypeSchema = async ({ body, di, set, params }: ElysiaContext) => {
     if (!params?.id) {
-        set.status = HTTP_STATUS.FORBIDDEN;
+        set.status = 400;
         return { success: false, data: 'Document type ID is required' };
     }
