+++ backend/packages/server/routes/handlers/handleGetDocumentTypesGroup.ts
@@ -5,7 +5,7 @@ import { resolveSchemaService } from "../utils/resolveSchemaService";
 
 export const handleGetDocumentTypesGroup = async ({ body, di, set, params }: ElysiaContext) => {
     if (!params?.id) {
-        set.status = HTTP_STATUS.FORBIDDEN;
+        set.status = 400;
         return { success: false, data: 'Group ID is required' };
     }
